mail_server/requirements.txt
python-dotenv==1.0.1
django==5.0.7
psycopg2-binary
psycopg
channels==4.1.0
daphne==4.1.2
channels-redis==4.2.0
djangorestframework==3.15.2
djoser==2.2.3
celery==5.4.0
redis==5.0.8
celery[redis]
eventlet==0.36.1